Réorganise les réductions d'impôt en plusieurs fichiers
