+++ openfisca_france/model/prelevements_obligatoires/impot_revenu/reductions_impot_plafonnees.py
@@ -19,14 +19,6 @@ log = logging.getLogger(__name__)
 # d'impôts, pour les déclarations après 2013, supérieures au plafond en vigueur l'année de la déclaration. Pour le moment on leur applique le montant en vigueur
 # l'année de la déclaration. L'approximation qui est faite est donc une approximation qui sousestime le montant global de ces réductions d'impôt.
 
-# Le cas particulier des investissements d'outremer (domsoc, domlog, doment): la mise à jour des formules et la prise en compte d'un plafond approximatif
-#   n'est réalisée qu'à partir de 2016. Au vu de la complexité de ces dispositifs (rétrocession, investissements avec des plafonds particuliers..),
-#   les plafonds sont approximés (par année et par dispositif) et ils ne prennent pas en compte les interactions entre les trois dispositifs d'outremer ni
-#   l'interaction avec le plafonnement global des avantages fiscaux. De ce point de vue, l'approximation surestime les plafonds pour les investissements d'outremer.
-#   D'un autre côté lors de la déclaration d'impôt il possible de choisir pour ces dispositifs entre un plafond absolu et un plafond relatif en pourcentage de RNI,
-#   seul le plafond absolu est codé pour le moment, cela peut entrainer une sous estimation des réductions d'impôt pour les personnes qui ont un RNI
-#   supérieur à environ 300 000.
-
 # TODO: le plafonnement global des réductions d'impôts avant 2013 (et la prise en compte du plafonnement des investissements d'outremer avant 2016)
 # TODO: La formule ci_invfor est à améliorer, l'ordre de priorité des variables est chronologique (en cas de dépassement du plafond, il
 # faut prendre en compte les variables les plus anciennes)
@@ -123,42 +115,6 @@ class reductions_plafonnees_esus_sfs(Variable):
         return red_esus_sfs
 
 
-class reductions_deplafonnees(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = "Réductions d'impôt sur le revenu déplafonnées"
-    definition_period = YEAR
-
-    def formula_2013_01_01(foyer_fiscal, period, parameters):
-
-        reductions_sans_plafond = [
-            'daepad',
-            'dfppce',
-            'adhcga',
-            'assvie',
-            'reduction_cotisations_syndicales',
-            'creaen',
-            'intagr',
-            'mecena',
-            'prcomp',
-            'repsoc',
-            'resimm',  # Malraux, non plafonnées pour les investissements réalisés après 2013
-            'ecpess',
-            'accult',
-            'rsceha',
-
-            # plafonds séparés, déjà appliqués dans les formules :
-            'doment',
-            'domlog',
-            'domsoc',
-            ]
-
-        # Step 4: Get other uncapped reductions
-        red_deplaf = sum([around(foyer_fiscal(reduction, period)) for reduction in reductions_sans_plafond])
-
-        return red_deplaf
-
-
 class reductions(Variable):
     value_type = float
     entity = FoyerFiscal
@@ -169,11 +125,12 @@ class reductions(Variable):
         '''
         Renvoie la somme des réductions d'impôt en prenant en compte les plafonds applicables.
 
-        Il y a 4 catégories de réductions :
+        Il y a 5 catégories de réductions :
             - réductions générales soumises au plafond de 10K
             - réductions inv. outre-mer/SOFICA avec majoration de 8K
             - réductions ESUS/SFS avec majoration de 3K (2021)
             - réductions sans plafond
+            - réductions d'outremer de la brochure 2042 IOM plafonnées avec des règles spécifiques supplémentaires
 
         Un tel plafond, avec un seuil différent, existe depuis 2008, mais n'est pour l'instant pas codé ici.
         La version codée là-dessous est celle de 2013 qui est encore valable pour l'imposition des revenus
@@ -195,8 +152,9 @@ class reductions(Variable):
         red_plaf_om_sofica = foyer_fiscal('reductions_plafonnees_om_sofica', period)
         red_plaf_esus_sfs = foyer_fiscal('reductions_plafonnees_esus_sfs', period)
         red_deplaf = foyer_fiscal('reductions_deplafonnees', period)
+        red_iom = foyer_fiscal('reductions_iom', period)
 
-        total_reduction = red_plaf + red_plaf_om_sofica + red_plaf_esus_sfs + red_deplaf
+        total_reduction = red_plaf + red_plaf_om_sofica + red_plaf_esus_sfs + red_deplaf + red_iom
 
         return min_(impot_net, total_reduction)
 
@@ -977,64 +935,6 @@ class duflot_pinel_denormandie_om(Variable):
         return reduction_cumulee + report + prorogation + f7sy
 
 
-class accult(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'Acquisition de biens culturels'
-    definition_period = YEAR
-    end = '2020-12-31'
-
-    def formula_2002(foyer_fiscal, period, parameters):
-        '''
-        Acquisition de biens culturels (case 7UO)
-        2002-
-        '''
-        f7uo = foyer_fiscal('f7uo', period)
-        P = parameters(period).impot_revenu.calcul_credits_impots.accult
-
-        return P.taux * f7uo
-
-
-class adhcga(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'adhcga'
-    definition_period = YEAR
-
-    def formula(foyer_fiscal, period, parameters):
-        '''
-        Frais de comptabilité et d'adhésion à un CGA ou AA
-        2002-
-        '''
-        f7ff = foyer_fiscal('f7ff', period)
-        f7fg = foyer_fiscal('f7fg', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.adhcga
-
-        return min_(f7ff, P.max * f7fg)
-
-
-class assvie(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'assvie'
-    end = '2004-12-31'
-    definition_period = YEAR
-
-    def formula_2002_01_01(foyer_fiscal, period, parameters):
-        '''
-        Assurance-vie (cases GW, GX et GY de la 2042)
-        2002-2004
-        '''
-        nb_pac = foyer_fiscal('nb_pac', period)
-        f7gw = foyer_fiscal('f7gw_2004', period)
-        f7gx = foyer_fiscal('f7gx_2004', period)
-        f7gy = foyer_fiscal('f7gy_2004', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.assvie
-
-        max1 = P.max + nb_pac * P.pac
-        return P.taux * min_(f7gw + f7gx + f7gy, max1)
-
-
 class cappme(Variable):
     value_type = float
     entity = FoyerFiscal
@@ -1691,118 +1591,6 @@ class cappme_esus_sfs(Variable):
         return base_esus_sfs
 
 
-class reduction_cotisations_syndicales(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = "Réduction d'impôt pour cotisations syndicales"
-    definition_period = YEAR
-    end = '2011-12-31'
-
-    def formula(foyer_fiscal, period, parameters):
-        '''
-        Cotisations syndicales : réduction d'impôt (2002-2011) puis crédit d'impôt (2012- )
-        '''
-        return foyer_fiscal('cotsyn', period)
-
-
-class cotsyn(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = "Montant de la réduction ou crédit d'impôt pour cotisations syndicales"
-    definition_period = YEAR
-
-    def formula(foyer_fiscal, period, parameters):
-        '''
-        NB: This is but an approximation of the true calculation, as it is done on the level of the tax unit.
-        In reality, the ceilings of 1 % of revenues are applied on the individual level (cf. BOI-IR-RICI-20).
-        '''
-
-        f7ac = foyer_fiscal.members('f7ac', period)
-
-        cotisations_versees = f7ac
-
-        salaire_imposable = foyer_fiscal.members('salaire_imposable', period, options = [ADD])
-        chomage_imposable = foyer_fiscal.members('chomage_imposable', period, options = [ADD])
-        retraite_imposable = foyer_fiscal.members('retraite_imposable', period, options = [ADD])
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.cotsyn
-
-        plafond = (salaire_imposable + chomage_imposable + retraite_imposable) * P.seuil
-
-        return (P.taux * foyer_fiscal.sum(min_(cotisations_versees, plafond)))
-
-
-class creaen(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'creaen'
-    definition_period = YEAR
-    end = '2014-12-31'
-
-    def formula_2006_01_01(foyer_fiscal, period, parameters):
-        '''
-        Aide aux créateurs et repreneurs d'entreprises
-        2006-2008
-        '''
-        f7fy = foyer_fiscal('f7fy_2011', period)
-        f7gy = foyer_fiscal('f7gy_2010', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.creaen
-
-        return (P.base * f7fy + P.hand * f7gy)
-
-    def formula_2009_01_01(foyer_fiscal, period, parameters):
-        '''
-        Aide aux créateurs et repreneurs d'entreprises
-        2009
-        '''
-        f7fy = foyer_fiscal('f7fy_2011', period)
-        f7gy = foyer_fiscal('f7gy_2010', period)
-        f7jy = foyer_fiscal('f7jy_2010', period)
-        f7hy = foyer_fiscal('f7hy_2011', period)
-        f7ky = foyer_fiscal('f7ky_2011', period)
-        f7iy = foyer_fiscal('f7iy_2011', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.creaen
-
-        return (
-            P.base * ((f7jy + f7fy) + f7hy / 2)
-            + P.hand * ((f7ky + f7gy) + f7iy / 2)
-            )
-
-    def formula_2010_01_01(foyer_fiscal, period, parameters):
-        '''
-        Aide aux créateurs et repreneurs d'entreprises
-        2010-2011
-        '''
-        f7fy = foyer_fiscal('f7fy_2011', period)
-        f7gy = foyer_fiscal('f7gy_2010', period)
-        f7jy = foyer_fiscal('f7jy_2010', period)
-        f7hy = foyer_fiscal('f7hy_2011', period)
-        f7ky = foyer_fiscal('f7ky_2011', period)
-        f7iy = foyer_fiscal('f7iy_2011', period)
-        f7ly = foyer_fiscal('f7ly_2010', period)
-        f7my = foyer_fiscal('f7my_2010', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.creaen
-
-        return (
-            P.base * ((f7jy + f7fy) + (f7hy + f7ly) / 2)
-            + P.hand * ((f7ky + f7gy) + (f7iy + f7my) / 2)
-            )
-
-    def formula_2012_01_01(foyer_fiscal, period, parameters):
-        '''
-        Aide aux créateurs et repreneurs d'entreprises
-        2012-
-        '''
-        f7ly = foyer_fiscal('f7ly_2010', period)
-        f7my = foyer_fiscal('f7my_2010', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.creaen
-
-        return (
-            P.base * (f7ly / 2)
-            + P.hand * (f7my / 2)
-            )
-
-
 class deffor(Variable):
     value_type = float
     entity = FoyerFiscal
@@ -1818,308 +1606,6 @@ class deffor(Variable):
         P = parameters(period).impot_revenu.calcul_reductions_impots.deffor
 
         return P.taux * min_(f7uc, P.max)
-
-
-class daepad(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'daepad'
-    definition_period = YEAR
-
-    def formula(foyer_fiscal, period, parameters):
-        '''
-        Dépenses d'accueil dans un établissement pour personnes âgées dépendantes
-        ?-
-        '''
-        f7cd = foyer_fiscal('f7cd', period)
-        f7ce = foyer_fiscal('f7ce', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.daepad
-
-        return P.taux * (min_(f7cd, P.max) + min_(f7ce, P.max))
-
-
-class dfppce(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = "Dons à des organismes d'intérêt général et dons pour le financement des partis politiques"
-    reference = 'http://bofip.impots.gouv.fr/bofip/5823-PGP'
-    definition_period = YEAR
-
-    def formula_2002_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons aux autres oeuvres et dons effectués pour le financement des partis
-        politiques et des campagnes électorales
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-
-        base = f7uf
-        max1 = P.max_dons_oeuvres * rni
-        return P.taux_dons_oeuvres * min_(base, max1)
-
-    def formula_2004_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons aux autres oeuvres et dons effectués pour le financement des partis
-        politiques et des campagnes électorales
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-
-        base = f7uf + f7xs
-        max1 = P.max_dons_oeuvres * rni
-        return P.taux_dons_oeuvres * min_(base, max1)
-
-    def formula_2005_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons aux autres oeuvres et dons effectués pour le financement des partis
-        politiques et des campagnes électorales
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-
-        base = f7uf + f7xs + f7xt
-        max1 = P.max_dons_oeuvres * rni
-        return P.taux_dons_oeuvres * min_(base, max1)
-
-    def formula_2006_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons aux autres oeuvres et dons effectués pour le financement des partis
-        politiques et des campagnes électorales
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-
-        base = f7uf + f7xs + f7xt + f7xu
-        max1 = P.max_dons_oeuvres * rni
-        return P.taux_dons_oeuvres * min_(base, max1)
-
-    def formula_2007_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons aux autres oeuvres et dons effectués pour le financement des partis
-        politiques et des campagnes électorales
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        f7xw = foyer_fiscal('f7xw', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-
-        base = f7uf + f7xs + f7xt + f7xu + f7xw
-        max1 = P.max_dons_oeuvres * rni
-        return P.taux_dons_oeuvres * min_(base, max1)
-
-    def formula_2008_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons aux autres oeuvres et dons effectués pour le financement des partis
-        politiques et des campagnes électorales
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        f7xw = foyer_fiscal('f7xw', period)
-        f7xy = foyer_fiscal('f7xy', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-
-        base = f7uf + f7xs + f7xt + f7xu + f7xw + f7xy
-        max1 = P.max_dons_oeuvres * rni
-        return P.taux_dons_oeuvres * min_(base, max1)
-
-    def formula_2011_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons aux autres oeuvres et dons effectués pour le financement des partis
-        politiques et des campagnes électorales (2011-2013)
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7ud = foyer_fiscal('f7ud', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        f7xw = foyer_fiscal('f7xw', period)
-        f7xy = foyer_fiscal('f7xy', period)
-        f7va = foyer_fiscal('f7va', period)
-        f7vc = foyer_fiscal('f7vc', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-        plafond_reduction_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.max
-
-        report_f7va_f7ud = max_(0, f7va + f7ud - plafond_reduction_donapd)
-        base = f7uf + f7vc + f7xs + f7xt + f7xu + f7xw + f7xy + report_f7va_f7ud
-        max1 = P.max_dons_oeuvres * rni
-        return P.taux_dons_oeuvres * min_(base, max1)
-
-    def formula_2012_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons aux autres oeuvres et dons effectués pour le financement des partis
-        politiques et des campagnes électorales (2011-2013)
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7ud = foyer_fiscal('f7ud', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        f7xw = foyer_fiscal('f7xw', period)
-        f7xy = foyer_fiscal('f7xy', period)
-        f7va = foyer_fiscal('f7va', period)
-        f7vc = foyer_fiscal('f7vc', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-        plafond_reduction_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.max
-
-        report_f7va_f7ud = max_(0, f7va + f7ud - plafond_reduction_donapd)
-        base = f7uf + f7vc + f7xs + f7xt + f7xu + f7xw + f7xy + report_f7va_f7ud
-        max1 = P.max_dons_oeuvres * rni
-        return P.taux_dons_oeuvres * min_(base, max1)
-
-    def formula_2013_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons versés à d’autres organismes d’intérêt général,
-        aux associations d’utilité publique, aux candidats aux élections (2019)
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7uh = foyer_fiscal('f7uh', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        f7xw = foyer_fiscal('f7xw', period)
-        f7xy = foyer_fiscal('f7xy', period)
-        f7va = foyer_fiscal('f7va', period)
-        f7ud = foyer_fiscal('f7ud', period)
-        f7vc = foyer_fiscal('f7vc', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-        plafond_reduction_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.max
-        taux_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.taux
-        maries_ou_pacses = foyer_fiscal('maries_ou_pacses', period)
-
-        red_7ud_7va = min_(plafond_reduction_donapd, f7va + f7ud) * taux_donapd
-        report_f7va_f7ud = max_(0, f7va + f7ud - plafond_reduction_donapd)
-
-        dons_partipol = min_(P.max_dons_partipo_seul * (1 + maries_ou_pacses), f7uh)
-
-        base = f7uf + f7vc + f7xs + f7xt + f7xu + f7xw + f7xy + report_f7va_f7ud + dons_partipol
-        max = P.max_dons_oeuvres * rni
-
-        return red_7ud_7va + P.taux_dons_oeuvres * min_(base, max)
-
-    def formula_2019_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons versés à d’autres organismes d’intérêt général,
-        aux associations d’utilité publique, aux candidats aux élections (2019)
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7uh = foyer_fiscal('f7uh', period)
-        f7ue = foyer_fiscal('f7ue', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        f7xw = foyer_fiscal('f7xw', period)
-        f7xy = foyer_fiscal('f7xy', period)
-        f7va = foyer_fiscal('f7va', period)
-        f7ud = foyer_fiscal('f7ud', period)
-        f7vc = foyer_fiscal('f7vc', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-        PND = parameters(period).impot_revenu.calcul_reductions_impots.dons.dons_notre_dame
-        plafond_reduction_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.max
-        taux_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.taux
-        maries_ou_pacses = foyer_fiscal('maries_ou_pacses', period)
-
-        red_7ud_7va = min_(plafond_reduction_donapd, f7va + f7ud) * taux_donapd
-        report_f7va_f7ud = max_(0, f7va + f7ud - plafond_reduction_donapd)
-
-        red_notre_dame = min_(PND.plafond, f7ue) * PND.taux
-        report_notre_dame = max_(0, f7ue - PND.plafond)
-
-        dons_partipol = min_(P.max_dons_partipo_seul * (1 + maries_ou_pacses), f7uh)
-
-        base = f7uf + f7vc + f7xs + f7xt + f7xu + f7xw + f7xy + report_f7va_f7ud + dons_partipol + report_notre_dame
-        max = P.max_dons_oeuvres * rni
-
-        return red_notre_dame + red_7ud_7va + P.taux_dons_oeuvres * min_(base, max)
-
-    def formula_2020_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons versés à d’autres organismes d’intérêt général,
-        aux associations d’utilité publique, aux candidats aux élections (2020)
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7uh = foyer_fiscal('f7uh', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        f7xw = foyer_fiscal('f7xw', period)
-        f7xy = foyer_fiscal('f7xy', period)
-        f7va = foyer_fiscal('f7va', period)
-        f7ud = foyer_fiscal('f7ud', period)
-        f7vc = foyer_fiscal('f7vc', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-        plafond_reduction_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.max
-        taux_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.taux
-        maries_ou_pacses = foyer_fiscal('maries_ou_pacses', period)
-
-        red_7ud_7va = min_(plafond_reduction_donapd, f7va + f7ud) * taux_donapd
-        report_f7va_f7ud = max_(0, f7va + f7ud - plafond_reduction_donapd)
-
-        dons_partipol = min_(P.max_dons_partipo_seul * (1 + maries_ou_pacses), f7uh)
-
-        base = f7uf + f7vc + f7xs + f7xt + f7xu + f7xw + f7xy + report_f7va_f7ud + dons_partipol
-        max = P.max_dons_oeuvres * rni
-
-        return red_7ud_7va + P.taux_dons_oeuvres * min_(base, max)
-
-    def formula_2021_01_01(foyer_fiscal, period, parameters):
-        '''
-        Dons versés à d’autres organismes d’intérêt général,
-        aux associations d’utilité publique, aux candidats aux élections (2021)
-        '''
-        rni = foyer_fiscal('rni', period)
-        f7uf = foyer_fiscal('f7uf', period)
-        f7uh = foyer_fiscal('f7uh', period)
-        f7uj = foyer_fiscal('f7uj', period)
-        f7xs = foyer_fiscal('f7xs', period)
-        f7xt = foyer_fiscal('f7xt', period)
-        f7xu = foyer_fiscal('f7xu', period)
-        f7xw = foyer_fiscal('f7xw', period)
-        f7xy = foyer_fiscal('f7xy', period)
-        f7va = foyer_fiscal('f7va', period)
-        f7ud = foyer_fiscal('f7ud', period)
-        f7vc = foyer_fiscal('f7vc', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.dons
-        plafond_reduction_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.max
-        taux_donapd = parameters(period).impot_revenu.calcul_reductions_impots.donapd.taux
-        maries_ou_pacses = foyer_fiscal('maries_ou_pacses', period)
-
-        red_7ud_7va = min_(plafond_reduction_donapd, f7va + f7ud) * taux_donapd
-        report_f7va_f7ud = max_(0, f7va + f7ud - plafond_reduction_donapd)
-
-        red_7uj = min_(P.dons_assoc_cult, f7uj) * taux_donapd
-        report_7uj = max_(0, f7uj - P.dons_assoc_cult)
-
-        dons_partipol = min_(P.max_dons_partipo_seul * (1 + maries_ou_pacses), f7uh)
-
-        base = f7uf + f7vc + f7xs + f7xt + f7xu + f7xw + f7xy + report_f7va_f7ud + report_7uj + dons_partipol
-        max = P.max_dons_oeuvres * rni
-
-        return red_7ud_7va + red_7uj + P.taux_dons_oeuvres * min_(base, max)
 
 
 class ecodev(Variable):
@@ -2139,31 +1625,6 @@ class ecodev(Variable):
         P = parameters(period).impot_revenu.calcul_reductions_impots.ecodev
 
         return min_(f7uh * P.taux, min_(P.taux_plafond * rbg_int, P.plafond_par_personne))  # page3 ligne 18
-
-
-class ecpess(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'ecpess'
-    definition_period = YEAR
-
-    def formula(foyer_fiscal, period, parameters):
-        '''
-        Réduction d'impôt au titre des enfants à charge poursuivant leurs études secondaires ou supérieures
-        '''
-        f7ea = foyer_fiscal('f7ea', period)
-        f7eb = foyer_fiscal('f7eb', period)
-        f7ec = foyer_fiscal('f7ec', period)
-        f7ed = foyer_fiscal('f7ed', period)
-        f7ef = foyer_fiscal('f7ef', period)
-        f7eg = foyer_fiscal('f7eg', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.ecpess
-
-        return (
-            P.col * (f7ea + f7eb / 2)
-            + P.lyc * (f7ec + f7ed / 2)
-            + P.sup * (f7ef + f7eg / 2)
-            )
 
 
 class garext(Variable):
@@ -2213,25 +1674,6 @@ class garext(Variable):
             + min_(f7gf, max2)
             + min_(f7gg, max2)
             )
-
-
-class intagr(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'intagr'
-    definition_period = YEAR
-
-    def formula_2005_01_01(foyer_fiscal, period, parameters):
-        '''
-        Intérêts pour paiement différé accordé aux agriculteurs
-        2005-
-        '''
-        f7um = foyer_fiscal('f7um', period)
-        maries_ou_pacses = foyer_fiscal('maries_ou_pacses', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.intagr
-
-        max1 = P.max * (1 + maries_ou_pacses)
-        return P.taux * min_(f7um, max1)
 
 
 class intcon(Variable):
@@ -3987,22 +3429,6 @@ class locmeu(Variable):
         return rep_ri + inv_ri
 
 
-class mecena(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = "Mécénat d'entreprise"
-    definition_period = YEAR
-
-    def formula_2003_01_01(foyer_fiscal, period, parameters):
-        '''
-        Mécénat d'entreprise (case 7US)
-        2003-
-        '''
-        f7us = foyer_fiscal('f7us', period)
-
-        return f7us
-
-
 class mohist(Variable):
     value_type = float
     entity = FoyerFiscal
@@ -4121,40 +3547,6 @@ class patnat(Variable):
         return f7ke
 
 
-class prcomp(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'Prestations compensatoires'
-    definition_period = YEAR
-
-    def formula(foyer_fiscal, period, parameters):
-        '''
-        Prestations compensatoires
-        '''
-        f7wm = foyer_fiscal('f7wm', period)
-        f7wn = foyer_fiscal('f7wn', period)
-        f7wo = foyer_fiscal('f7wo', period)
-        f7wp = foyer_fiscal('f7wp', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.prcomp
-
-        div = (f7wo == 0) * 1 + f7wo  # Pour éviter les divisions par zéro
-
-        return (
-            (f7wm == 0) * (
-                (f7wn == f7wo) * P.taux * min_(f7wn, P.seuil)
-                + (f7wn < f7wo) * (f7wo <= P.seuil) * P.taux * f7wn
-                + max_(0, (f7wn < f7wo) * (f7wo > P.seuil) * P.taux * P.seuil * f7wn / div)
-                )
-            + (f7wm != 0) * (
-                (f7wn == f7wm) * (f7wo <= P.seuil) * P.taux * f7wm
-                + max_(0, (f7wn == f7wm) * (f7wo >= P.seuil) * P.taux * f7wm / div)
-                + (f7wn > f7wm) * (f7wo <= P.seuil) * P.taux * f7wn
-                + max_(0, (f7wn > f7wm) * (f7wo >= P.seuil) * P.taux * f7wn / div)
-                )
-            + P.taux * f7wp
-            )
-
-
 class reduction_impot_exceptionnelle(Variable):
     value_type = float
     entity = FoyerFiscal
@@ -4190,318 +3582,6 @@ class rehab(Variable):
         depenses_2017 = min_(P.max, f7xx)
 
         return (P.taux * depenses_2017)
-
-
-class repsoc(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'repsoc'
-    definition_period = YEAR
-
-    def formula_2003_01_01(foyer_fiscal, period, parameters):
-        '''
-        Intérèts d'emprunts pour reprises de société
-        2003-
-        '''
-        maries_ou_pacses = foyer_fiscal('maries_ou_pacses', period)
-        f7fh = foyer_fiscal('f7fh', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.repsoc
-
-        seuil = P.seuil * (maries_ou_pacses + 1)
-        return P.taux * min_(f7fh, seuil)
-
-
-class resimm(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = "Réduction d'impôt au titre des dépenses de restauration immobilière - Dispositif Malraux"
-    reference = 'http://bofip.impots.gouv.fr/bofip/1372-PGP'
-    definition_period = YEAR
-
-    def formula_2009_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière (cases 7RA et 7RB)
-        2009-2010
-        '''
-        f7ra = foyer_fiscal('f7ra_2015', period)
-        f7rb = foyer_fiscal('f7rb_2015', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        max1 = P.max
-        max2 = max_(max1 - f7rb, 0)
-        return P.taux_rb * min_(f7rb, max1) + P.taux_ra * min_(f7ra, max2)
-
-    def formula_2011_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière (cases 7RA, 7RB, 7RC, 7RD)
-        2011
-        '''
-        f7ra = foyer_fiscal('f7ra_2015', period)
-        f7rb = foyer_fiscal('f7rb_2015', period)
-        f7rc = foyer_fiscal('f7rc_2015', period)
-        f7rd = foyer_fiscal('f7rd_2015', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        max1 = P.max
-        max2 = max_(max1 - f7rd, 0)
-        max3 = max_(max2 - f7rb, 0)
-        max4 = max_(max3 - f7rc, 0)
-
-        return (
-            P.taux_rd * min_(f7rd, max1)
-            + P.taux_rb * min_(f7rb, max2)
-            + P.taux_rc * min_(f7rc, max3)
-            + P.taux_ra * min_(f7ra, max4)
-            )
-
-    def formula_2012_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière (cases 7RA, 7RB, 7RC, 7RD, 7RE, 7RF)
-        2012
-        '''
-        f7ra = foyer_fiscal('f7ra_2015', period)
-        f7rb = foyer_fiscal('f7rb_2015', period)
-        f7rc = foyer_fiscal('f7rc_2015', period)
-        f7rd = foyer_fiscal('f7rd_2015', period)
-        f7re = foyer_fiscal('f7re_2016', period)
-        f7rf = foyer_fiscal('f7rf_2016', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        max1 = P.max
-        max2 = max_(max1 - f7rd, 0)
-        max3 = max_(max2 - f7rb, 0)
-        max4 = max_(max3 - f7rc - f7rf, 0)
-        max5 = max_(max4 - f7ra, 0)
-
-        return (
-            P.taux_rd * min_(f7rd, max1)
-            + P.taux_rb * min_(f7rb, max2)
-            + P.taux_rc * min_(f7rc + f7rf, max3)
-            + P.taux_ra * min_(f7ra, max4)
-            + P.taux_re * min_(f7re, max5)
-            )
-
-    def formula_2013_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière (cases 7RA, 7RB, 7RC, 7RD, 7RE, 7RF, 7SX, 7SY)
-        2013-2015
-        '''
-        f7ra = foyer_fiscal('f7ra_2015', period)
-        f7rb = foyer_fiscal('f7rb_2015', period)
-        f7rc = foyer_fiscal('f7rc_2015', period)
-        f7rd = foyer_fiscal('f7rd_2015', period)
-        f7re = foyer_fiscal('f7re_2016', period)
-        f7rf = foyer_fiscal('f7rf_2016', period)
-        f7sx = foyer_fiscal('f7sx_2017', period)
-        f7sy = foyer_fiscal('f7sy_2017', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        max1 = P.max
-        max2 = max_(max1 - f7rd, 0)
-        max3 = max_(max2 - f7rb, 0)
-        max4 = max_(max3 - f7rc - f7sy - f7rf, 0)
-        max5 = max_(max4 - f7ra, 0)
-
-        return (
-            P.taux_rd * min_(f7rd, max1)
-            + P.taux_rb * min_(f7rb, max2)
-            + P.taux_rc * min_(f7sy + f7rf + f7rc, max3)
-            + P.taux_ra * min_(f7ra, max4) + P.taux_re * min_(f7re + f7sx, max5)
-            )
-
-    def formula_2016_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière
-        2016
-        '''
-        f7nx = foyer_fiscal('f7nx', period)
-        f7ny = foyer_fiscal('f7ny', period)
-        f7re = foyer_fiscal('f7re_2016', period)
-        f7rf = foyer_fiscal('f7rf_2016', period)
-        f7sx = foyer_fiscal('f7sx_2017', period)
-        f7sy = foyer_fiscal('f7sy_2017', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        max1 = P.max
-        max2 = max_(max1 - f7nx - f7sy - f7rf, 0)
-
-        return (
-            P.taux_rc * min_(f7sy + f7rf + f7nx, max1)
-            + P.taux_re * min_(f7re + f7sx + f7ny, max2)
-            )
-
-    def formula_2017_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière
-        2017
-        '''
-        f7sy = foyer_fiscal('f7sy_2017', period)
-        f7sx = foyer_fiscal('f7sx_2017', period)
-
-        f7nx = foyer_fiscal('f7nx', period)
-        f7ny = foyer_fiscal('f7ny', period)
-
-        f7tx = foyer_fiscal('f7tx', period)
-        f7ty = foyer_fiscal('f7ty', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        depenses_secteur_degrade = f7sy + f7nx
-        depenses_secteur_patrimoine_remarquable = f7sx + f7ny
-        depenses_PSMV_2017 = f7tx
-        depenses_non_PSMV_2017 = f7ty
-
-        max1 = max_(P.max - depenses_secteur_degrade, 0)
-        max3 = max_(P.max2 - depenses_PSMV_2017, 0)
-
-        return (
-            P.taux_30 * (
-                min_(depenses_secteur_degrade, P.max)
-                + min_(depenses_PSMV_2017, P.max2)
-                )
-            + P.taux_22 * (
-                min_(depenses_secteur_patrimoine_remarquable, max1)
-                + min_(depenses_non_PSMV_2017, max3)
-                )
-            )
-
-    def formula_2018_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière
-        2018
-        '''
-        # plaf 100K
-        f7nx = foyer_fiscal('f7nx', period)
-        f7ny = foyer_fiscal('f7ny', period)
-
-        # plaf 400K
-        f7tx = foyer_fiscal('f7tx', period)
-        f7ty = foyer_fiscal('f7ty', period)
-
-        # reports
-        f7kz = foyer_fiscal('f7kz', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        base_nx = min_(P.max, f7nx)
-        base_ny = min_(P.max - f7nx, f7ny)
-
-        base_tx = min_(P.max2, f7tx)
-        base_ty = min_(P.max2 - f7tx, f7ty)
-
-        ri = (f7kz
-            + P.taux_30 * (base_nx + base_tx)
-            + P.taux_22 * (base_ny + base_ty))
-
-        return ri
-
-    def formula_2019_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière
-        2019
-        '''
-        # plaf 100K
-        f7nx = foyer_fiscal('f7nx', period)
-        f7ny = foyer_fiscal('f7ny', period)
-
-        # plaf 400K
-        f7tx = foyer_fiscal('f7tx', period)
-        f7ty = foyer_fiscal('f7ty', period)
-
-        # reports
-        f7kz = foyer_fiscal('f7kz', period)
-        f7ky = foyer_fiscal('f7ky', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        base_nx = min_(P.max, f7nx)
-        base_ny = min_(P.max - f7nx, f7ny)
-
-        base_tx = min_(P.max2, f7tx)
-        base_ty = min_(P.max2 - f7tx, f7ty)
-
-        ri = (f7kz + f7ky
-            + P.taux_30 * (base_nx + base_tx)
-            + P.taux_22 * (base_ny + base_ty))
-
-        return ri
-
-    def formula_2020_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière
-        2020
-        '''
-        # plaf 100K
-        f7nx = foyer_fiscal('f7nx', period)
-        f7ny = foyer_fiscal('f7ny', period)
-
-        # plaf 400K
-        f7tx = foyer_fiscal('f7tx', period)
-        f7ty = foyer_fiscal('f7ty', period)
-
-        # reports
-        f7kz = foyer_fiscal('f7kz', period)
-        f7ky = foyer_fiscal('f7ky', period)
-        f7kx = foyer_fiscal('f7kx', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        base_nx = min_(P.max, f7nx)
-        base_ny = min_(P.max - f7nx, f7ny)
-
-        base_tx = min_(P.max2, f7tx)
-        base_ty = min_(P.max2 - f7tx, f7ty)
-
-        ri = (f7kz + f7ky + f7kx
-            + P.taux_30 * (base_nx + base_tx)
-            + P.taux_22 * (base_ny + base_ty))
-
-        return ri
-
-    def formula_2021_01_01(foyer_fiscal, period, parameters):
-        '''
-        Travaux de restauration immobilière
-        2021
-        '''
-        # plaf 400K
-        f7tx = foyer_fiscal('f7tx', period)
-        f7ty = foyer_fiscal('f7ty', period)
-
-        # reports
-        f7ky = foyer_fiscal('f7ky', period)
-        f7kx = foyer_fiscal('f7kx', period)
-        f7kw = foyer_fiscal('f7kw', period)
-
-        P = parameters(period).impot_revenu.calcul_reductions_impots.resimm
-
-        base_tx = min_(P.max2, f7tx)
-        base_ty = min_(P.max2 - f7tx, f7ty)
-
-        ri = (f7ky + f7kx + f7kw
-            + P.taux_30 * (base_tx)
-            + P.taux_22 * (base_ty))
-
-        return ri
-
-
-class rsceha(Variable):
-    value_type = float
-    entity = FoyerFiscal
-    label = 'rsceha'
-    definition_period = YEAR
-
-    def formula(foyer_fiscal, period, parameters):
-        '''
-        Rentes de survie et contrats d'épargne handicap
-        2002-
-        '''
-        nb_pac_majoration_plafond = foyer_fiscal('nb_pac2', period)
-        nbR = foyer_fiscal('nbR', period)
-        f7gz = foyer_fiscal('f7gz', period)
-        P = parameters(period).impot_revenu.calcul_reductions_impots.rsceha
-
-        max1 = P.seuil1 + (nb_pac_majoration_plafond - nbR) * P.seuil2
-        return P.taux * min_(f7gz, max1)
 
 
 class ri_saldom(Variable):
